Fix Visual Studio warnings

+++ src/yakc/systems/atom.cc
@@ -164,7 +164,7 @@ atom_t::reset() {
 uint64_t
 atom_t::exec(uint64_t start_tick, uint64_t end_tick) {
     YAKC_ASSERT(start_tick <= end_tick);
-    uint32_t num_ticks = end_tick - start_tick;
+    uint32_t num_ticks = uint32_t(end_tick - start_tick);
     uint32_t ticks_executed = m6502_exec(&board.m6502, num_ticks);
     /* did we hit the osload() trap? */
     if (board.m6502.trap_id == 1) {
