+++ src/yakc/systems/z9001.cc
@@ -170,7 +170,7 @@ z9001_t::reset() {
 uint64_t
 z9001_t::exec(uint64_t start_tick, uint64_t end_tick) {
     YAKC_ASSERT(start_tick <= end_tick);
-    uint32_t num_ticks = end_tick - start_tick;
+    uint32_t num_ticks = uint32_t(end_tick - start_tick);
     uint32_t ticks_executed = z80_exec(&board.z80, num_ticks);
     kbd_update(&board.kbd);
     this->decode_video();
